requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.1
starlette==0.37.2
pydantic==2.8.2
aiohttp==3.9.5
httpx==0.27.0
prometheus_client==0.20.0
PyYAML==6.0.1
motor==3.5.1
qdrant-client==1.9.0
typer==0.12.3
redis==5.0.7

# OpenAI Integration
openai==1.45.0
tiktoken==0.7.0

# Technical Analysis & Strategy Development
ta-lib==0.4.28
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

# Plugin System
importlib-metadata==8.0.0
pluggy==1.5.0

# Performance & Monitoring
psutil==6.0.0
